Made the PixelButton component more dynamic and merged the two animations into one component
