--- src/Common/Components/Buttons/PixelButtonFadeOut/PixelButtonFadeOut.scss
@@ -1,57 +1,0 @@
-@use '../../../scss/_colors.ts' as *;
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-.pixel-button-container-FO {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 2vh;
-}
-
-.pixel-button-FO {
-    position: relative;
-    width: var(--width);
-    height: var(--height);
-    border: none;
-    background: $neutral-darkest;
-    color: $anti-flash-white;
-    font-size: 1.15em;
-    text-transform:  uppercase;
-    letter-spacing: 0.1em;
-    cursor: pointer;
-    outline: none;
-}
-
-.pixel-button-FO span {
-    position: relative;
-    z-index: 2;
-}
-
-.pixel-container-FO {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    z-index: 1;
-    pointer-events: none;
-}
-
-.pixel-container-FO .pixel-FO {
-    position: absolute;
-    width: 10px;
-    height: 10px;
-    pointer-events: none;
-    background: var(--color);
-    transition: transform 0.5s ease, opacity 0.5s ease;
-}
-
-.pixel-button-FO:hover .pixel-FO {
-    opacity: 0;
-    transform: translate(var(--tx), var(--ty));
-}
